mcu/nrf5340: Fix virtual flash driver compilation
Add dedicated IPC API for getting embedded netcore image information
instead of vflash relying on internals of IPC driver.
