+++ hw/mcu/nordic/nrf5340_net/src/hal_vflash.c
@@ -27,16 +27,9 @@
 #include <nrfx_ipc.h>
 #include <bootutil/bootutil.h>
 #include <bootutil/image.h>
-#include <ipc_nrf5340/ipc_nrf5340_priv.h>
+#include <ipc_nrf5340/ipc_nrf5340.h>
 
 #define NRF5340_NET_VFLASH_SECTOR_SZ 2048
-
-#if !MYNEWT_VAL(IPC_NRF5340_PRE_TRUSTZONE_NETCORE_BOOT)
-__attribute__((section(".ipc"))) static struct ipc_shared ipc_shared[1];
-#endif
-
-#define NRF_APP_IPC_NS                  ((NRF_IPC_Type *)0x4002A000)
-#define NRF_APP_IPC_S                   ((NRF_IPC_Type *)0x5002A000)
 
 struct swap_data {
     uint32_t encryption_key0[4];
@@ -222,13 +215,10 @@ static int
 nrf5340_net_vflash_init(const struct hal_flash *dev)
 {
     struct nrf5340_vflash *vflash = (struct nrf5340_vflash *)dev;
-#if MYNEWT_VAL(IPC_NRF5340_PRE_TRUSTZONE_NETCORE_BOOT)
-    const void *img_addr = (const void *)NRF_APP_IPC_S->GPMEM[0];
-    uint32_t image_size = (uint32_t)NRF_APP_IPC_S->GPMEM[1];
-#else
-    const void *img_addr = ipc_shared[0].net_core_image_address;
-    uint32_t image_size = ipc_shared[0].net_core_image_size;
-#endif
+    const void *img_addr;
+    uint32_t image_size;
+
+    img_addr = ipc_nrf5340_net_image_get(&image_size);
 
     /*
      * Application side IPC will set ipc_share data
